hw6

+++ src/main/java/ru/gb/homework3/controller/ProductController.java
@@ -1,11 +1,10 @@
 package ru.gb.homework3.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import ru.gb.homework3.model.Product;
+import ru.gb.homework3.entity.Product;
 import ru.gb.homework3.service.ProductService;
 
 @Controller
@@ -15,7 +14,6 @@ public class ProductController {
 
     private ProductService productService;
 
-
     @RequestMapping(value = "/add", method = RequestMethod.GET)
     public String showForm(Model model) {
         model.addAttribute("product", new Product());
@@ -23,7 +21,7 @@ public class ProductController {
     }
 
     @RequestMapping(path = "/{id}", method = RequestMethod.GET)
-    public String getProductId(Model model, @PathVariable Integer id) {
+    public String getProductId(Model model, @PathVariable Long id) {
         Product product;
         product = productService.findById(id);
         model.addAttribute("product", product);
